move all settings into the same `settings.py` module
